% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_num_common_cons.R
\name{calc_num_common_conns}
\alias{calc_num_common_conns}
\title{Title}
\usage{
calc_num_common_conns(member, f_of_f, member_conns)
}
\arguments{
\item{member_conns}{}
}
\value{
something
}
\description{
Title
}
