% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_fof_igraph_to_df.R
\name{convert_fof_igraph_to_df}
\alias{convert_fof_igraph_to_df}
\title{Convert Friend of Friend igraph Object to Data Frame}
\usage{
convert_fof_igraph_to_df(igraph_connections, friends_of_friends_vertices)
}
\arguments{
\item{friends_of_friends_vertices}{}
}
\value{
a data frame
}
\description{
Convert Friend of Friend igraph Object to Data Frame
}
